+++ CHANGELOG.md
@@ -1,10 +1,46 @@
+# [1.2.0-dev.4](https://github.com/m-xim/fluentus/compare/v1.2.0-dev.3...v1.2.0-dev.4) (2025-01-09)
+
+
+### Bug Fixes
+
+* **fluent_api:** replace `fluent.runtime` error handling with ParseError from `fluent.syntax` ([51b2be3](https://github.com/m-xim/fluentus/commit/51b2be3a3ef9014872fa2b36dd24459bcfc498d0))
+
+# [1.2.0-dev.3](https://github.com/m-xim/fluentus/compare/v1.2.0-dev.2...v1.2.0-dev.3) (2025-01-09)
+
+
+### Bug Fixes
+
+* **fluent_api:** enable uploading of files from subdirectories ([d25ad11](https://github.com/m-xim/fluentus/commit/d25ad11cd54f66dbb4391ee2c10a5998ad2c47f9))
+
+# [1.2.0-dev.2](https://github.com/m-xim/fluentus/compare/v1.2.0-dev.1...v1.2.0-dev.2) (2025-01-09)
+
+
+### Bug Fixes
+
+* **table_manager:** change dict to defaultdict ([4798cdd](https://github.com/m-xim/fluentus/commit/4798cddfdda5e7e5857566dab3da6ed8f39552e1))
+
+# [1.2.0-dev.1](https://github.com/m-xim/fluentus/compare/v1.1.1-dev.3...v1.2.0-dev.1) (2025-01-07)
+
+
+### Features
+
+* **table:** display all translations in the table ([41e2670](https://github.com/m-xim/fluentus/commit/41e26708bbdc5260984a8dca855b7c565a3676e1))
+
+## [1.1.1-dev.3](https://github.com/m-xim/fluentus/compare/v1.1.1-dev.2...v1.1.1-dev.3) (2025-01-07)
+
+
+### Bug Fixes
+
+* **api:** ensure translations are properly cleared ([1514859](https://github.com/m-xim/fluentus/commit/1514859fe6b7be5affd08874f85c6159f7b514b6))
+* **editor:** cursor position ([2f8f078](https://github.com/m-xim/fluentus/commit/2f8f078735e05b9f19ef48df33d0bd3354468b67))
+* **fluent_api:** fix whitespace (\n,\s,\t,...) characters at end of text, refactor code and add translation types ([9de7e6d](https://github.com/m-xim/fluentus/commit/9de7e6d12889ab43db7d4383599e40b71b67c1fb))
+
 ## [1.1.1](https://github.com/m-xim/fluentus/compare/v1.1.0...v1.1.1) (2025-01-05)
 
 
 ### Bug Fixes
 
-* **editor:** cursor position ([2f8f078](https://github.com/m-xim/fluentus/commit/2f8f078735e05b9f19ef48df33d0bd3354468b67))
-* **fluent_api:** fix whitespace (\n,\s,\t,...) characters at end of text, refactor code and add translation types ([9de7e6d](https://github.com/m-xim/fluentus/commit/9de7e6d12889ab43db7d4383599e40b71b67c1fb))
+* **fluent_api:** enable uploading of files from subdirectories ([d25ad11](https://github.com/m-xim/fluentus/commit/d25ad11cd54f66dbb4391ee2c10a5998ad2c47f9))
 
 ## [1.1.1-dev.2](https://github.com/m-xim/fluentus/compare/v1.1.1-dev.1...v1.1.1-dev.2) (2025-01-05)